refactor(audit-log): move audit log utility functions to a separate file for better organization
